requirements.txt
django==1.8.3
whitenoise
django-toolbelt
pillow
djrill
django-cors-headers
python-telegram-bot